refactor: remove unnecessary custom exceptions

+++ src/main/java/com/swifty/bank/server/core/domain/customer/service/impl/CustomerServiceImpl.java
@@ -1,22 +1,17 @@
 package com.swifty.bank.server.core.domain.customer.service.impl;
 
-import com.swifty.bank.server.core.domain.customer.dto.CustomerInfoResponse;
-import com.swifty.bank.server.core.domain.customer.dto.CustomerInfoUpdateConditionRequest;
 import com.swifty.bank.server.core.domain.customer.Customer;
 import com.swifty.bank.server.core.domain.customer.constant.CustomerStatus;
+import com.swifty.bank.server.core.domain.customer.dto.CustomerInfoResponse;
+import com.swifty.bank.server.core.domain.customer.dto.CustomerInfoUpdateConditionRequest;
 import com.swifty.bank.server.core.domain.customer.dto.JoinRequest;
-import com.swifty.bank.server.core.domain.customer.exceptions.CannotReferCustomerByNullException;
-import com.swifty.bank.server.core.domain.customer.exceptions.NoSuchCustomerByDeviceID;
-import com.swifty.bank.server.core.domain.customer.exceptions.NoSuchCustomerByPhoneNumberException;
-import com.swifty.bank.server.core.domain.customer.exceptions.NoSuchCustomerByUUID;
 import com.swifty.bank.server.core.domain.customer.repository.CustomerRepository;
 import com.swifty.bank.server.core.domain.customer.service.CustomerService;
-import java.util.Objects;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 
-import java.util.NoSuchElementException;
+import java.util.Objects;
 import java.util.UUID;
 
 @Transactional(readOnly = true)
@@ -36,6 +31,9 @@ public class CustomerServiceImpl implements CustomerService {
                 .phoneNumber(joinRequest.getPhoneNumber())
                 .password(joinRequest.getPassword())
                 .deviceId(joinRequest.getDeviceId())
+                .birthDate(joinRequest.getBirthDate())
+                .gender(joinRequest.getGender())
+                .roles(joinRequest.getRoles())
                 .build();
 
         customerRepository.save(customer);
@@ -48,35 +46,32 @@ public class CustomerServiceImpl implements CustomerService {
     // Send access token(JWT) to frontend with encrypted UUID
     // Condition of Retrieval : JPQL
     @Override
-    public Customer findByUuid(UUID uuid) throws CannotReferCustomerByNullException {
+    public Customer findByUuid(UUID uuid) {
         if (uuid == null) {
-            throw new CannotReferCustomerByNullException();
+            return null;
         }
 
         return customerRepository.findOneByUUID(uuid)
-                .orElseThrow(() -> new NoSuchElementException("Customer not found"));
+                .orElse(null);
     }
 
     @Override
-    public Customer findByDeviceId(String deviceId) throws CannotReferCustomerByNullException {
+    public Customer findByDeviceId(String deviceId) {
         if (deviceId == null)
-            throw new CannotReferCustomerByNullException();
+            return null;
 
         return customerRepository.findOneByDeviceId(deviceId)
-                .orElseThrow(() -> new NoSuchCustomerByDeviceID("[ERROR] No result as referring with device ID"));
+                .orElse(null);
     }
 
     @Override
-    public Customer findByPhoneNumber(String phoneNumber) throws CannotReferCustomerByNullException {
+    public Customer findByPhoneNumber(String phoneNumber) {
         if (phoneNumber == null) {
-            throw new CannotReferCustomerByNullException();
+            return null;
         }
 
         return customerRepository.findOneByPhoneNumber(phoneNumber)
-                .orElseThrow(() ->
-                        new NoSuchCustomerByPhoneNumberException("[ERROR] No Result as referring with phone " +
-                                "number and nationality")
-                );
+                .orElse(null);
     }
 
     @Transactional
@@ -84,15 +79,19 @@ public class CustomerServiceImpl implements CustomerService {
     public Customer updateCustomerInfo(UUID customerUuid, CustomerInfoUpdateConditionRequest customerInfoUpdateConditionRequest) {
 
         Customer customer = customerRepository.findOneByUUID(customerUuid)
-                .orElseThrow(() -> new NoSuchElementException("회원이 존재하지 않습니다"));
+                .orElse(null);
 
-        if(Objects.nonNull(customerInfoUpdateConditionRequest.getName())) customer.updateName(customerInfoUpdateConditionRequest.getName());
+        if (Objects.nonNull(customerInfoUpdateConditionRequest.getName()))
+            customer.updateName(customerInfoUpdateConditionRequest.getName());
 
-        if(Objects.nonNull(customerInfoUpdateConditionRequest.getPhoneNumber())) customer.updatePhoneNumber(customerInfoUpdateConditionRequest.getPhoneNumber());
+        if (Objects.nonNull(customerInfoUpdateConditionRequest.getPhoneNumber()))
+            customer.updatePhoneNumber(customerInfoUpdateConditionRequest.getPhoneNumber());
 
-        if(Objects.nonNull(customerInfoUpdateConditionRequest.getBirthDate())) customer.updateBirthDate(customerInfoUpdateConditionRequest.getBirthDate());
+        if (Objects.nonNull(customerInfoUpdateConditionRequest.getBirthDate()))
+            customer.updateBirthDate(customerInfoUpdateConditionRequest.getBirthDate());
 
-        if(Objects.nonNull(customerInfoUpdateConditionRequest.getNationality())) customer.updateNationality(customerInfoUpdateConditionRequest.getNationality());
+        if (Objects.nonNull(customerInfoUpdateConditionRequest.getNationality()))
+            customer.updateNationality(customerInfoUpdateConditionRequest.getNationality());
 
         return customer;
     }
@@ -101,9 +100,7 @@ public class CustomerServiceImpl implements CustomerService {
     @Override
     public Customer updatePhoneNumber(UUID uuid, String phoneNumber) {
         Customer customer = customerRepository.findOneByUUID(uuid)
-                .orElseThrow(() -> new NoSuchCustomerByPhoneNumberException("[ERROR] No customer " +
-                        "found by the phone" +
-                        " number and nationality"));
+                .orElse(null);
 
         customer.updatePhoneNumber(phoneNumber);
         return customer;
@@ -113,7 +110,7 @@ public class CustomerServiceImpl implements CustomerService {
     @Override
     public Customer updateDeviceId(UUID uuid, String deviceId) {
         Customer customer = customerRepository.findOneByUUID(uuid)
-                .orElseThrow(() -> new NoSuchCustomerByUUID("[ERROR] : No customer found by the device id"));
+                .orElse(null);
 
         customer.updateDeviceId(deviceId);
         return customer;
@@ -121,17 +118,22 @@ public class CustomerServiceImpl implements CustomerService {
 
     @Transactional
     @Override
-    public void withdrawCustomer(UUID uuid) {
+    public Customer withdrawCustomer(UUID uuid) {
         Customer customer = customerRepository.findOneByUUID(uuid)
-                .orElseThrow(() -> new NoSuchCustomerByUUID("No such Customer"));
+                .orElse(null);
+
+        if (customer == null) {
+            return null;
+        }
 
         customer.delete();
+        return customer;
     }
 
     @Override
     public CustomerInfoResponse findCustomerInfoDtoByUuid(UUID uuid) {
         CustomerInfoResponse customerInfoResponse = customerRepository.findCustomerInfoResponseByUUID(uuid)
-                .orElseThrow(() -> new NoSuchElementException("No such Customer"));
+                .orElse(null);
 
         return customerInfoResponse;
     }
@@ -139,7 +141,7 @@ public class CustomerServiceImpl implements CustomerService {
     @Override
     public void updatePassword(UUID uuid, String newPassword) {
         Customer customer = customerRepository.findOneByUUID(uuid)
-                .orElseThrow(() -> new NoSuchElementException("No such Customer"));
+                .orElse(null);
 
         customer.resetPassword(newPassword);
     }
